% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\docType{data}
\name{population}
\alias{population}
\title{Total population statistics in South African provinces}
\format{A data frame with 45 rows and 5 variables:
\describe{
 \item{province}{One of the nine provinces of South Africa: Eastern Cape,
 Free State, Gauteng, KwaZulu-Natal, Limpopo, Mpumalanga, North West,
 Northern Cape, or Western Cape.}
 \item{longitude}{Longitude of the province, as reported by Google Maps.}
 \item{latitude}{Latitude of the province, as reported by Google Maps.}
 \item{year}{The year, from 2011 to 2016.}
 \item{total}{Total population in the given province and year.}
}}
\source{
\url{http://southafrica.opendataforafrica.org/qrglcbb/population-statistics-of-south-africa-2011}
}
\usage{
population
}
\description{
A dataset from Statistics South Africa on total population in South African
provinces. Total population for all provinces in South Africa is reported
for the years 2011 to 2016.
}
\keyword{datasets}

